installed react, deleted files not needed, import useState, useRef, and useEffect, jsx code in return ()
renderCount is an object with a current property which ='s itself with 1 added to it everytime react renders. if name is changed or when you move the space cursor the page renders and the count goes up. The important thing about Ref is that you can store a previous value in it & it persists between different renders but it does not cause you to rerender like useState does. That's one of the reason for using useRef the other is referencing elements